requirements.txt
django
django-environ
django-cors-headers
psycopg2
djangorestframework
djangorestframework_simplejwt
Pillow
djoser